+++ src/main/java/blackjack/model/user/Dealer.java
@@ -1,7 +1,7 @@
 package blackjack.model.user;
 
 import blackjack.model.constant.Constant;
-import blackjack.model.GameResult;
+import blackjack.model.gameresult.GameResult;
 import java.util.EnumMap;
 
 public class Dealer extends User {
