+++ src/main/java/blackjack/model/user/User.java
@@ -1,7 +1,7 @@
 package blackjack.model.user;
 
 import blackjack.model.card.Card;
-import blackjack.model.Hand;
+import blackjack.model.card.Hand;
 import java.util.List;
 
 public abstract class User {
