+++ src/test/java/blackjack/model/CardProviderTest.java
@@ -2,6 +2,7 @@ package blackjack.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.model.card.CardProvider;
 import blackjack.model.user.Dealer;
 import blackjack.model.user.Player;
 import blackjack.model.user.Users;
