+++ src/test/java/blackjack/model/HandTest.java
@@ -3,6 +3,7 @@ package blackjack.model;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.model.card.Card;
+import blackjack.model.card.Hand;
 import blackjack.model.card.Rank;
 import blackjack.model.card.Suit;
 import org.junit.jupiter.api.DisplayName;
