+++ src/test/java/blackjack/model/HitCommandTest.java
@@ -5,6 +5,7 @@ import static blackjack.model.constant.ErrorMessage.ERROR_NOT_Y_N_INPUT;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
+import blackjack.model.card.HitCommand;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
