+++ src/test/java/blackjack/util/PlayerParserTest.java
@@ -3,6 +3,7 @@ package blackjack.util;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.model.user.Player;
+import blackjack.model.user.PlayerParser;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
